Clean up more cruft
